requirements.txt
requests
python-dotenv
spotipy==2.25.1
beautifulsoup4==4.13.5
youtube-search-python==1.6.6
yt-dlp==2025.9.5
html5lib==1.1
